Refactor: memberInfoResponse 로 변경 및 패키지 이동
