+++ src/main/java/com/challenger/fridge/dto/member/MemberInfoResponse.java
@@ -1,4 +1,4 @@
-package com.challenger.fridge.dto;
+package com.challenger.fridge.dto.member;
 
 import com.challenger.fridge.domain.Member;
 import com.challenger.fridge.domain.Storage;
@@ -6,12 +6,11 @@ import com.challenger.fridge.dto.box.response.StorageBoxNameResponse;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class MemberInfoDto {
+public class MemberInfoResponse {
 
     private String username;
     private String email;
@@ -19,7 +18,7 @@ public class MemberInfoDto {
     private String mainStorageName;
     private List<StorageBoxNameResponse> storageBoxes;
 
-    public MemberInfoDto(Member member) {
+    public MemberInfoResponse(Member member) {
         this.username = member.getName();
         this.email = member.getEmail();
         Storage storage = member.getStorageList().get(0);
